refactor: update API endpoints and mode explore data source to it's package
