+++ data/src/main/java/com/di/RemoteSourceModule.kt
@@ -1,6 +1,6 @@
 package com.di
 
-import com.remote.ExploreRemoteDataSource
+import com.remote.source.ExploreRemoteDataSource
 import com.remote.source.SearchRemoteDataSource
 import io.ktor.client.engine.okhttp.OkHttp
 import org.koin.dsl.module
