+++ data/src/main/java/com/remote/source/ExploreRemoteDataSource.kt
@@ -1,22 +1,22 @@
-package com.remote
+package com.remote.source
 
 import com.remote.dto.GenreDto
 import com.remote.dto.GenreResponse
 import com.remote.dto.MovieDetailsDto
 import com.remote.dto.MovieDto
-import com.remote.dto.PopularResponse
-import com.remote.dto.SeriesDetailsDto
 import com.remote.dto.SeriesDto
 import com.utils.ApiResponse
 import com.utils.DISCOVER_MOVIE_LIST
 import com.utils.DISCOVER_SERIES_LIST
 import com.utils.GENRE_MOVIE_LIST
 import com.utils.GENRE_SERIES_LIST
+import com.utils.MOVIE
+import com.utils.PAGE
+import com.utils.POPULAR
+import com.utils.SERIES
 import com.utils.WITH_GENRES
 import com.utils.performCall
 import io.ktor.client.HttpClient
-import io.ktor.client.call.body
-import io.ktor.client.request.get
 import io.ktor.client.request.parameter
 import io.ktor.http.HttpMethod
 
@@ -25,38 +25,49 @@ class ExploreRemoteDataSource(
 ) {
     suspend fun getMoviesGenres(): List<GenreDto> =
         client.performCall<Unit, GenreResponse>(
-            method = HttpMethod.Get,
+            method = HttpMethod.Companion.Get,
             path = GENRE_MOVIE_LIST
         ).genres
 
     suspend fun getSeriesGenres() =
         client.performCall<Unit, GenreResponse>(
-            method = HttpMethod.Get,
+            method = HttpMethod.Companion.Get,
             path = GENRE_SERIES_LIST
         ).genres
 
-    suspend fun getMovies(): PopularResponse<MovieDto> =
-        client.get("https://api.themoviedb.org/3/movie/popular") {
-            parameter("language", "en-US")
-            parameter("page", 1)
-        }.body()
-    suspend fun getMovieDetails(id: Int): MovieDetailsDto =
-        client.get("https://api.themoviedb.org/3/movie/$id") {
-            parameter("language", "en-US")
-        }.body()
+    suspend fun getMovies(): List<MovieDto> =
+        client.performCall<Unit, ApiResponse<MovieDto>>(
+            method = HttpMethod.Get,
+            path = MOVIE + POPULAR
+        ) {
+            parameter(PAGE, 1)
+        }.results
 
-    suspend fun getSeries(): PopularResponse<SeriesDto> =
-        client.get("https://api.themoviedb.org/3/tv/popular") {
-            parameter("language", "en-US")
-            parameter("page", 1)
-        }.body()
-    suspend fun getSeriesDetails(id: Int): SeriesDetailsDto =
-        client.get("https://api.themoviedb.org/3/tv/$id") {
-            parameter("language", "en-US")
-        }.body()
-    suspend fun getSeriesByGenreId(genreId: Int): List<SeriesDto> =
+
+    suspend fun getSeries(): List<SeriesDto> =
         client.performCall<Unit, ApiResponse<SeriesDto>>(
             method = HttpMethod.Get,
+            path = SERIES + POPULAR
+        ) {
+            parameter(PAGE, 1)
+        }.results
+
+    suspend fun getMovieDetails(id: Int): MovieDetailsDto =
+        client.performCall<Unit, MovieDetailsDto>(
+            method = HttpMethod.Get,
+            path = MOVIE + id
+        )
+
+
+    suspend fun getSeriesDetails(id: Int): SeriesDto =
+        client.performCall<Unit, SeriesDto>(
+            method = HttpMethod.Get,
+            path = SERIES + id
+        )
+
+    suspend fun getSeriesByGenreId(genreId: Int): List<SeriesDto> =
+        client.performCall<Unit, ApiResponse<SeriesDto>>(
+            method = HttpMethod.Companion.Get,
             path = DISCOVER_SERIES_LIST,
             requestBuilder = {
                 parameter(WITH_GENRES, genreId)
@@ -65,7 +76,7 @@ class ExploreRemoteDataSource(
 
     suspend fun getMoviesByGenreId(genreId: Int): List<MovieDto> =
         client.performCall<Unit, ApiResponse<MovieDto>>(
-            method = HttpMethod.Get,
+            method = HttpMethod.Companion.Get,
             path = DISCOVER_MOVIE_LIST,
             requestBuilder = {
                 parameter(WITH_GENRES, genreId)
