+++ data/src/main/java/com/utils/Endpoints.kt
@@ -11,3 +11,7 @@ const val GENRE_SERIES_LIST ="genre/tv/list"
 const val DISCOVER ="discover/"
 const val DISCOVER_MOVIE_LIST ="${DISCOVER}movie"
 const val DISCOVER_SERIES_LIST ="${DISCOVER}tv"
+const val MOVIE = "movie/"
+const val SERIES = "tv/"
+const val ACTOR = "person/"
+const val POPULAR = "popular"
