+++ local.properties
@@ -8,4 +8,4 @@
 # For customization when using a Version Control System, please read the
 # header note.
 sdk.dir=C\:\\Users\\LEGION\\AppData\\Local\\Android\\Sdk
-TMDB_API_KEY=
+TMDB_API_KEY=b37d9f568685efadafa2ce0a871597fb
